+++ payments/models.py
@@ -1,4 +1,4 @@
-import json
+import warnings
 from typing import Iterable
 from typing import Union
 from uuid import uuid4
@@ -12,29 +12,6 @@ from . import FraudStatus
 from . import PaymentStatus
 from . import PurchasedItem
 from .core import provider_factory
-
-
-class PaymentAttributeProxy:
-    def __init__(self, payment):
-        self._payment = payment
-        super().__init__()
-
-    def __getattr__(self, item):
-        data = json.loads(self._payment.extra_data or "{}")
-        try:
-            return data[item]
-        except KeyError as e:
-            raise AttributeError(*e.args)
-
-    def __setattr__(self, key, value):
-        if key == "_payment":
-            return super().__setattr__(key, value)
-        try:
-            data = json.loads(self._payment.extra_data)
-        except ValueError:
-            data = {}
-        data[key] = value
-        self._payment.extra_data = json.dumps(data)
 
 
 class BasePayment(models.Model):
@@ -78,7 +55,7 @@ class BasePayment(models.Model):
     billing_email = models.EmailField(blank=True)
     billing_phone = PhoneNumberField(blank=True)
     customer_ip_address = models.GenericIPAddressField(blank=True, null=True)
-    extra_data = models.TextField(blank=True, default="")
+    extra_data = models.JsonField(blank=True, default=dict)
     message = models.TextField(blank=True, default="")
     token = models.CharField(max_length=36, blank=True, default="")
     captured_amount = models.DecimalField(max_digits=9, decimal_places=2, default="0.0")
@@ -195,14 +172,18 @@ class BasePayment(models.Model):
 
     @property
     def attrs(self):
-        """A JSON-serialised wrapper around `extra_data`.
+        warnings.warn(
+            "Using BasePayment.attrs is deprecated. Use BasePayment.extra_data instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return self.extra_data
 
-        This property exposes a a dict or list which is serialised into the `extra_data`
-        text field. Usage of this wrapper is preferred over accessing the underlying
-        field directly.
-
-        You may think of this as a `JSONField` which is saved to the `extra_data`
-        column.
-        """
-        # TODO: Deprecate in favour of JSONField when we drop support for django 2.2.
-        return PaymentAttributeProxy(self)
+    @attrs.setter
+    def attrs(self, value):
+        warnings.warn(
+            "Using BasePayment.attrs is deprecated. Use BasePayment.extra_data instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        self.extra_data = value